Move database queue creation into SecureStorage.
